compose/arrivalinfo 패키지 arrivalinfo로 합병
