+++ app/src/main/java/com/hansung/sherpa/ui/searchscreen/ExpandableCard.kt
@@ -42,7 +42,7 @@ import com.hansung.sherpa.R
 import com.hansung.sherpa.SherpaScreen
 import com.hansung.sherpa.StaticValue
 import com.hansung.sherpa.arrivalinfo.ODsayArrivalInfoRequest
-import com.hansung.sherpa.compose.busarrivalinfo.ArrivalInfoManager
+import com.hansung.sherpa.arrivalinfo.ArrivalInfoManager
 import com.hansung.sherpa.compose.chart.Chart
 import com.hansung.sherpa.itemsetting.BusLane
 import com.hansung.sherpa.itemsetting.BusSectionInfo
